cobalt/requirements.txt
django==1.9.4


# Forms
django-braces==1.8.1
django-crispy-forms==1.6.0
django-floppyforms==1.6.1

#images
Pillow==3.1.1